[fitting] New concept for neighborhood frames
