+++ cmake/PoncaConfigureFitting.cmake
@@ -18,6 +18,7 @@ set(ponca_Fitting_INCLUDE
     "${PONCA_src_ROOT}/Ponca/src/Fitting/dryFit.h"
     "${PONCA_src_ROOT}/Ponca/src/Fitting/enums.h"
     "${PONCA_src_ROOT}/Ponca/src/Fitting/evaluationScheme.h"
+    "${PONCA_src_ROOT}/Ponca/src/Fitting/frame.h"
     "${PONCA_src_ROOT}/Ponca/src/Fitting/gls.h"
     "${PONCA_src_ROOT}/Ponca/src/Fitting/gls.hpp"
     "${PONCA_src_ROOT}/Ponca/src/Fitting/heightField.h"
